code review: move code into sub-packages
