+++ pkg/xds/envoy/clusters/clusters_suite_test.go
@@ -1,0 +1,13 @@
+package clusters_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestClusters(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Clusters Suite")
+}
