+++ pkg/xds/envoy/clusters/clusters_test.go
@@ -1,47 +1,27 @@
-package envoy_test
+package clusters_test
 
 import (
 	"time"
-
-	"github.com/Kong/kuma/pkg/core/xds"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 
-	mesh_proto "github.com/Kong/kuma/api/mesh/v1alpha1"
-	mesh_core "github.com/Kong/kuma/pkg/core/resources/apis/mesh"
-	util_proto "github.com/Kong/kuma/pkg/util/proto"
-	xds_context "github.com/Kong/kuma/pkg/xds/context"
-	"github.com/Kong/kuma/pkg/xds/envoy"
+	. "github.com/Kong/kuma/pkg/xds/envoy/clusters"
+
+	"github.com/golang/protobuf/ptypes"
 
 	envoy_v2 "github.com/envoyproxy/go-control-plane/envoy/api/v2"
-	"github.com/golang/protobuf/ptypes"
+
+	mesh_proto "github.com/Kong/kuma/api/mesh/v1alpha1"
+	mesh_core "github.com/Kong/kuma/pkg/core/resources/apis/mesh"
+	core_xds "github.com/Kong/kuma/pkg/core/xds"
+	xds_context "github.com/Kong/kuma/pkg/xds/context"
+
+	util_proto "github.com/Kong/kuma/pkg/util/proto"
 )
 
-var _ = Describe("Envoy", func() {
-
-	It("should generate 'static' Endpoints", func() {
-		// given
-		expected := `
-        clusterName: localhost:8080
-        endpoints:
-        - lbEndpoints:
-          - endpoint:
-              address:
-                socketAddress:
-                  address: 127.0.0.1
-                  portValue: 8080
-`
-		// when
-		resource := envoy.CreateStaticEndpoint("localhost:8080", "127.0.0.1", 8080)
-
-		// then
-		actual, err := util_proto.ToYAML(resource)
-
-		Expect(err).ToNot(HaveOccurred())
-		Expect(actual).To(MatchYAML(expected))
-	})
+var _ = Describe("Clusters", func() {
 
 	It("should generate 'local' Cluster", func() {
 		// given
@@ -61,7 +41,7 @@ var _ = Describe("Envoy", func() {
                     portValue: 8080
 `
 		// when
-		resource := envoy.CreateLocalCluster("localhost:8080", "127.0.0.1", 8080)
+		resource := CreateLocalCluster("localhost:8080", "127.0.0.1", 8080)
 
 		// then
 		actual, err := util_proto.ToYAML(resource)
@@ -79,7 +59,7 @@ var _ = Describe("Envoy", func() {
         connectTimeout: 5s
 `
 		// when
-		resource := envoy.CreatePassThroughCluster("pass_through")
+		resource := CreatePassThroughCluster("pass_through")
 
 		// then
 		actual, err := util_proto.ToYAML(resource)
@@ -88,75 +68,18 @@ var _ = Describe("Envoy", func() {
 		Expect(actual).To(MatchYAML(expected))
 	})
 
-	Describe("CreateLbMetadata()", func() {
-
-		It("should handle `nil` map of tags", func() {
-			// when
-			metadata := envoy.CreateLbMetadata(nil)
-			// then
-			Expect(metadata).To(BeNil())
-		})
-
-		It("should handle empty map of tags", func() {
-			// when
-			metadata := envoy.CreateLbMetadata(map[string]string{})
-			// then
-			Expect(metadata).To(BeNil())
-		})
-
-		type testCase struct {
-			tags     map[string]string
-			expected string
-		}
-		DescribeTable("should generate Envoy metadata",
-			func(given testCase) {
-				// when
-				metadata := envoy.CreateLbMetadata(given.tags)
-				// and
-				actual, err := util_proto.ToYAML(metadata)
-				// then
-				Expect(err).ToNot(HaveOccurred())
-				Expect(actual).To(MatchYAML(given.expected))
-			},
-			Entry("map with 1 tag", testCase{
-				tags: map[string]string{
-					"service": "redis",
-				},
-				expected: `
-        filterMetadata:
-          envoy.lb:
-            service: redis
-`,
-			}),
-			Entry("map with multiple tags", testCase{
-				tags: map[string]string{
-					"service": "redis",
-					"version": "v1",
-					"region":  "eu",
-				},
-				expected: `
-        filterMetadata:
-          envoy.lb:
-            service: redis
-            version: v1
-            region: eu
-`,
-			}),
-		)
-	})
-
 	Describe("'EDS' Cluster", func() {
 
 		type testCase struct {
 			ctx      xds_context.Context
-			metadata xds.DataplaneMetadata
+			metadata core_xds.DataplaneMetadata
 			expected string
 		}
 
 		DescribeTable("should generate 'EDS' Cluster",
 			func(given testCase) {
 				// when
-				resource := envoy.CreateEdsCluster(given.ctx, "192.168.0.1:8080", &given.metadata)
+				resource := CreateEdsCluster(given.ctx, "192.168.0.1:8080", &given.metadata)
 
 				// then
 				actual, err := util_proto.ToYAML(resource)
@@ -171,7 +94,7 @@ var _ = Describe("Envoy", func() {
 						Resource: &mesh_core.MeshResource{},
 					},
 				},
-				metadata: xds.DataplaneMetadata{},
+				metadata: core_xds.DataplaneMetadata{},
 				expected: `
                 connectTimeout: 5s
                 edsClusterConfig:
@@ -198,7 +121,7 @@ var _ = Describe("Envoy", func() {
 						},
 					},
 				},
-				metadata: xds.DataplaneMetadata{},
+				metadata: core_xds.DataplaneMetadata{},
 				expected: `
                 connectTimeout: 5s
                 edsClusterConfig:
@@ -253,7 +176,7 @@ var _ = Describe("Envoy", func() {
 						},
 					},
 				},
-				metadata: xds.DataplaneMetadata{
+				metadata: core_xds.DataplaneMetadata{
 					DataplaneTokenPath: "/var/secret/token",
 				},
 				expected: `
@@ -313,80 +236,6 @@ var _ = Describe("Envoy", func() {
 		)
 	})
 
-	Describe("ClusterLoadAssignment()", func() {
-		type testCase struct {
-			cluster   string
-			endpoints []xds.Endpoint
-			expected  string
-		}
-		DescribeTable("should generate ClusterLoadAssignment",
-			func(given testCase) {
-				// when
-				resource := envoy.CreateClusterLoadAssignment(given.cluster, given.endpoints)
-
-				// then
-				actual, err := util_proto.ToYAML(resource)
-
-				Expect(err).ToNot(HaveOccurred())
-				Expect(actual).To(MatchYAML(given.expected))
-			},
-			Entry("without tags", testCase{
-				cluster: "127.0.0.1:8080",
-				endpoints: []xds.Endpoint{
-					{Target: "192.168.0.1", Port: 8081},
-					{Target: "192.168.0.2", Port: 8082},
-				},
-				expected: `
-        clusterName: 127.0.0.1:8080
-        endpoints:
-        - lbEndpoints:
-          - endpoint:
-              address:
-                socketAddress:
-                  address: 192.168.0.1
-                  portValue: 8081
-          - endpoint:
-              address:
-                socketAddress:
-                  address: 192.168.0.2
-                  portValue: 8082
-`,
-			}),
-			Entry("with tags", testCase{
-				cluster: "127.0.0.1:8080",
-				endpoints: []xds.Endpoint{
-					{Target: "192.168.0.1", Port: 8081, Tags: map[string]string{"service": "backend", "region": "us"}},
-					{Target: "192.168.0.2", Port: 8082, Tags: map[string]string{"service": "backend", "region": "eu"}},
-				},
-				expected: `
-        clusterName: 127.0.0.1:8080
-        endpoints:
-        - lbEndpoints:
-          - endpoint:
-              address:
-                socketAddress:
-                  address: 192.168.0.1
-                  portValue: 8081
-            metadata:
-              filterMetadata:
-                envoy.lb:
-                  region: us
-                  service: backend
-          - endpoint:
-              address:
-                socketAddress:
-                  address: 192.168.0.2
-                  portValue: 8082
-            metadata:
-              filterMetadata:
-                envoy.lb:
-                  region: eu
-                  service: backend
-`,
-			}),
-		)
-	})
-
 	Describe("ClusterWithHealthChecks()", func() {
 
 		type testCase struct {
@@ -400,7 +249,7 @@ var _ = Describe("Envoy", func() {
 					Name: "example",
 				}
 				// when
-				metadata := envoy.ClusterWithHealthChecks(cluster, given.healthCheck)
+				metadata := ClusterWithHealthChecks(cluster, given.healthCheck)
 				// and
 				actual, err := util_proto.ToYAML(metadata)
 				// then
@@ -510,4 +359,5 @@ var _ = Describe("Envoy", func() {
 			}),
 		)
 	})
+
 })
