+++ pkg/xds/envoy/endpoints/endpoints_suite_test.go
@@ -1,0 +1,13 @@
+package endpoints_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestEndpoints(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Endpoints Suite")
+}
