+++ pkg/xds/envoy/listeners/listener_suite_test.go
@@ -1,0 +1,13 @@
+package listeners_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestListener(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Listener Suite")
+}
