+++ pkg/xds/envoy/listeners/network_access_log_configurer.go
@@ -1,4 +1,4 @@
-package envoy
+package listeners
 
 import (
 	"fmt"
