+++ pkg/xds/envoy/listeners/network_rbac_configurer.go
@@ -1,4 +1,4 @@
-package envoy
+package listeners
 
 import (
 	"fmt"
