+++ pkg/xds/envoy/listeners/prometheus_endpoint_configurer.go
@@ -1,4 +1,4 @@
-package envoy
+package listeners
 
 import (
 	"github.com/golang/protobuf/ptypes"
