+++ pkg/xds/envoy/listeners/prometheus_endpoint_configurer_test.go
@@ -1,12 +1,13 @@
-package envoy_test
+package listeners_test
 
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 
+	. "github.com/Kong/kuma/pkg/xds/envoy/listeners"
+
 	util_proto "github.com/Kong/kuma/pkg/util/proto"
-	. "github.com/Kong/kuma/pkg/xds/envoy"
 )
 
 var _ = Describe("PrometheusEndpointConfigurer", func() {
