+++ pkg/xds/envoy/tls_test.go
@@ -1,0 +1,10 @@
+package envoy_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	_ "github.com/onsi/gomega"
+)
+
+var _ = Describe("Envoy", func() {
+
+})
